+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add plugin: renderfindings - Render single finding to PDF
 * Update and optimize designs: demo, HTB and OffSec designs
 * Add design for HTB CAPE certification
+* Warn about relative URLs in PDFs during rendering
 
 
 ## v2024.96 - 2024-12-04